Reorganize the plugin API
