+++ betty/content_provider/__init__.py
@@ -10,10 +10,10 @@ from typing import TYPE_CHECKING, final
 from betty.locale.localizable import _
 from betty.plugin import (
     ClassedPluginDefinition,
-    EntryPointDiscovery,
     HumanFacingPluginDefinition,
     PluginTypeDefinition,
 )
+from betty.plugin.discovery.entry_point import EntryPointDiscovery
 
 if TYPE_CHECKING:
     from betty.resource import Context
