+++ betty/license/__init__.py
@@ -10,14 +10,14 @@ from typing import TYPE_CHECKING, ClassVar, final
 from betty.locale.localizable import _
 from betty.mutability import Mutable
 from betty.plugin import (
-    AppDiscovery,
     ClassedPlugin,
     ClassedPluginDefinition,
-    EntryPointDiscovery,
     HumanFacingPluginDefinition,
     PluginTypeDefinition,
-    ProjectDiscovery,
 )
+from betty.plugin.discovery.app import AppDiscovery
+from betty.plugin.discovery.entry_point import EntryPointDiscovery
+from betty.plugin.discovery.project import ProjectDiscovery
 
 if TYPE_CHECKING:
     from betty.locale.localizable import Localizable
