+++ betty/tests/plugin/config/test___init__.py
@@ -13,7 +13,6 @@ from betty.machine_name import MachineName
 from betty.plugin import PluginRepository
 from betty.plugin.config import (
     HumanFacingPluginDefinitionConfiguration,
-    OrderedPluginDefinitionConfiguration,
     PluginDefinitionConfiguration,
     PluginIdentifierKeyConfigurationMapping,
     PluginInstanceConfiguration,
@@ -198,61 +197,6 @@ class TestHumanFacingPluginDefinitionConfiguration:
             id=plugin_id, label="", description=init_description
         )
         assert sut.description[expected_locale] == expected_description
-
-
-class TestOrderedPluginDefinitionConfiguration:
-    async def test_load__minimal(self) -> None:
-        dump: Dump = {"id": "hello-world"}
-        sut = OrderedPluginDefinitionConfiguration(id="-")
-        sut.load(dump)
-        assert not sut.comes_before
-        assert not sut.comes_after
-
-    async def test_load__with_comes_before(self) -> None:
-        dump: Dump = {
-            "id": "hello-world",
-            "comes_before": ["my-first-plugin"],
-        }
-        sut = OrderedPluginDefinitionConfiguration(id="-")
-        sut.load(dump)
-        assert sut.comes_before == {"my-first-plugin"}
-
-    async def test_load__with_comes_after(self) -> None:
-        dump: Dump = {
-            "id": "hello-world",
-            "comes_after": ["my-first-plugin"],
-        }
-        sut = OrderedPluginDefinitionConfiguration(id="-")
-        sut.load(dump)
-        assert sut.comes_after == {"my-first-plugin"}
-
-    async def test_dump__minimal(self) -> None:
-        sut = OrderedPluginDefinitionConfiguration(id="-")
-        dump = sut.dump()
-        assert "comes_before" not in dump
-        assert "comes_after" not in dump
-
-    async def test_dump__with_comes_before(self) -> None:
-        comes_before = {"my-first-plugin"}
-        sut = OrderedPluginDefinitionConfiguration(id="-", comes_before=comes_before)
-        dump = sut.dump()
-        assert dump["comes_before"] == list(comes_before)
-
-    async def test_dump__with_comes_after(self) -> None:
-        comes_after = {"my-first-plugin"}
-        sut = OrderedPluginDefinitionConfiguration(id="-", comes_after=comes_after)
-        dump = sut.dump()
-        assert dump["comes_after"] == list(comes_after)
-
-    async def test_comes_before(self) -> None:
-        comes_before = {"my-first-plugin"}
-        sut = OrderedPluginDefinitionConfiguration(id="-", comes_before=comes_before)
-        assert sut.comes_before == comes_before
-
-    async def test_comes_after(self) -> None:
-        comes_after = {"my-first-plugin"}
-        sut = OrderedPluginDefinitionConfiguration(id="-", comes_after=comes_after)
-        assert sut.comes_after == comes_after
 
 
 class TestPluginInstanceConfiguration:
