+++ betty/tests/plugin/discovery/test_entry_point.py
@@ -1,0 +1,37 @@
+from __future__ import annotations
+
+from importlib.metadata import EntryPoint, EntryPoints
+from typing import TYPE_CHECKING
+
+from betty.plugin.discovery.entry_point import EntryPointDiscovery
+from betty.test_utils.plugin import ClassedDummyPluginOne, ClassedDummyPluginTwo
+
+if TYPE_CHECKING:
+    from pytest_mock import MockerFixture
+
+
+class TestEntryPointDiscovery:
+    async def test_discover(self, mocker: MockerFixture) -> None:
+        entry_point_group = "test-entry-point"
+        m_entry_points = mocker.patch(
+            "importlib.metadata.entry_points",
+            return_value=EntryPoints(
+                [
+                    EntryPoint(
+                        name=ClassedDummyPluginOne.plugin.id,
+                        value="betty.test_utils.plugin:ClassedDummyPluginOne",
+                        group=entry_point_group,
+                    ),
+                    EntryPoint(
+                        name=ClassedDummyPluginTwo.plugin.id,
+                        value="betty.test_utils.plugin:ClassedDummyPluginTwo",
+                        group=entry_point_group,
+                    ),
+                ]
+            ),
+        )
+        sut = EntryPointDiscovery(entry_point_group)
+        plugins = await sut.discover(None)
+        assert ClassedDummyPluginOne.plugin in plugins
+        assert ClassedDummyPluginTwo.plugin in plugins
+        m_entry_points.assert_called_once_with(group=entry_point_group)
